+++ geolocator/test/geolocator_test.dart
@@ -18,8 +18,9 @@ Position get mockPosition => Position(
     speedAccuracy: 0.0);
 
 NmeaMessage get mockNmeaMessage => NmeaMessage(
-    "GPGGA,170834,4124.8963,N,08151.6838,W,1,05,1.5,280.2,M,-34.0,M,,,*75",
-    DateTime.fromMillisecondsSinceEpoch(
+    message:
+        "GPGGA,170834,4124.8963,N,08151.6838,W,1,05,1.5,280.2,M,-34.0,M,,,*75",
+    timestamp: DateTime.fromMillisecondsSinceEpoch(
       500,
       isUtc: true,
     ));
@@ -104,8 +105,8 @@ void main() {
       expect(position, emitsInOrder([emits(mockPosition), emitsDone]));
     });
 
-    test('getNmeaMessageStream', () async {
-      final nmeaMessage = Geolocator.getNmeaMessageStream();
+    test('getNmeaStream', () async {
+      final nmeaMessage = Geolocator.getNmeaStream();
       expect(nmeaMessage, emitsInOrder([emits(mockNmeaMessage), emitsDone]));
     });
 
@@ -190,7 +191,7 @@ class MockGeolocatorPlatform extends Mock
   }
 
   @override
-  Stream<NmeaMessage> getNmeaMessageStream() => Stream.value(mockNmeaMessage);
+  Stream<NmeaMessage> getNmeaStream() => Stream.value(mockNmeaMessage);
 
   @override
   Future<bool> openAppSettings() => Future.value(true);