+++ geolocator_platform_interface/test/geolocator_platform_interface_test.dart
@@ -160,14 +160,14 @@ void main() {
 
     test(
         // ignore: lines_longer_than_80_chars
-        'Default implementation of getNmeaMessageStream should throw unimplemented error',
+        'Default implementation of getNmeaStream should throw unimplemented error',
         () {
       // Arrange
       final geolocatorPlatform = ExtendsGeolocatorPlatform();
 
       // Act & Assert
       expect(
-        geolocatorPlatform.getNmeaMessageStream,
+        geolocatorPlatform.getNmeaStream,
         throwsUnimplementedError,
       );
     });