+++ geolocator_platform_interface/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 - Added functionality to receive NMEA messages.
 
+## 4.0.5
+
+- Fixes repository URL of the package.
+
+## 4.0.4
+
+- Fixes a bug where listening to the position stream immediately after an error, results in listening to a dead stream.
+
 ## 4.0.3
 
 - Removes `timeInterval` from `LocationSettings` documentation.