+++ geolocator_windows/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+- Fixes repository URL of the package.
+
 ## 0.1.0
 
 - Adds an initial implementation of Windows support for the geolocator plugin.